refactor: subqueries
Instead of wrapping the query in a special object, add a private '_alias' field, which can be set
with the 'as' or 'alias' methods. The '_from' and '_joins' properties only accept strings now, and
subqueries are parsed upon insertion in either property, rather than later at the build step.
Overall, this simplifies both the code and the user experience.

BREAKING CHANGE: QueryAs type and object no longer exist. No longer accepts subqueries in the 'joinRaw' method.