fix: refactor Query.clone() to copy all enumerable properties
Query.clone() now copies all enumerable properties instead of using a pre-defined list. This change simplifies the method and allows copying properties that are not defined in the initializer. It also fixes the bug where 'with' clauses didn't get copied and avoids future bugs of the same type. Previously, it was necessary to include new properties in the 'PROPERTY_TO_CONDITION' mapping, which was easy to forget, and doing so wouldn't trigger any errors. With this change, if the developer forgets to add a new property to the 'queryPropertyNamesAndDefaultValues' mapping, it still gets cloned properly, but it can't be cleared (in the 'clone' or 'clear' methods). Trying to do so will throw an error both at compile and run time. Also, cloning is used more often than clearing.